+++ src/text_symbolizer.cpp
@@ -87,7 +87,7 @@ IMPLEMENT_ENUM( text_transform_e, text_transform_strings )
 
 
 text_symbolizer::text_symbolizer(expression_ptr name, std::string const& face_name,
-                                 unsigned size, color const& fill,
+                                 float size, color const& fill,
                                  text_placements_ptr placements)
     : symbolizer_base(),
       name_(name),
@@ -120,7 +120,7 @@ text_symbolizer::text_symbolizer(expression_ptr name, std::string const& face_na
     set_text_size(size);
 }
 
-text_symbolizer::text_symbolizer(expression_ptr name, unsigned size, color const& fill,
+text_symbolizer::text_symbolizer(expression_ptr name, float size, color const& fill,
                                  text_placements_ptr placements)
     : symbolizer_base(),
       name_(name),
@@ -378,12 +378,12 @@ void text_symbolizer::set_max_char_angle_delta(double angle)
     max_char_angle_delta_ = angle;
 }
 
-void text_symbolizer::set_text_size(unsigned size)
+void text_symbolizer::set_text_size(float size)
 {
     placement_options_->set_default_text_size(size);
 }
 
-unsigned  text_symbolizer::get_text_size() const
+float text_symbolizer::get_text_size() const
 {
     return placement_options_->get_default_text_size();
 }
